pyproject.toml
[build-system]
requires = [
    "poetry>=1.0.0",
]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "exasol-bucketfs"
packages  = [
    {include = "exasol"},
    {include = "exasol_bucketfs_utils_python"}
]
version = "0.13.0"
description = "BucketFS utilities for the Python programming language"

license = "MIT"

authors = [
    "Torsten Kilias <torsten.kilias@exasol.com>",
    "Nicola Coretti <nicola.coretti@exasol.com>",
    "Anastasiia Sergienko <anastasiia.sergienko@exasol.com>"
]

readme = 'README.rst'

repository = "https://github.com/exasol/bucketfs-python"
homepage = "https://github.com/exasol/bucketfs-python"

keywords = ['exasol', 'bucketfs']

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
requests = ">=2.24.0"
joblib=">=1.0.1"
typeguard = ">=4.3.0"
exasol-saas-api=">=0.3.0"
# Temp.fix
httpx=">=0.27.0"
attrs=">=23.2.0"

[tool.poetry.dev-dependencies]
pyexasol = "^0.25.2"
dill = "^0.3.4"
exasol-udf-mock-python = ">=0.2.0"
toml = ">=0.10.2"
sphinx-copybutton = "^0.5.0"
exasol-toolbox = ">=0.16.0,<1"
pytest-localserver = "^0.8.1"
pytest-exasol-backend = ">=0.3.2,<1"

[tool.coverage.run]
source = [
    "exasol",
]

[tool.coverage.report]
fail_under = 15


[tool.black]
line-length = 88
verbose = false
include = "\\.pyi?$"


[tool.isort]
profile = "black"
force_grid_wrap = 2


[tool.pylint.master]
fail-under = 7.0
output-format = "colorized,json:.lint.json,text:.lint.txt"

[tool.pylint.format]
max-line-length = 88
max-module-lines = 800

[tool.mypy]
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "test.*",
]
ignore_errors = true
